Fixed not loading api which doesn't have defined spider. Fixed fullscreen not showing loading hint. Fixed fullscreen button on web remote control. Fixed IJK ffmpeg issue. Fixed menu button to pop Source switch dialog.
