+++ .gitignore
@@ -1,6 +1,7 @@
 # Built application files
 *.apk
 *.aar
+!/app/libs/*.aar
 *.ap_
 *.aab
