+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 2
-        versionName '0.8.20220909.1'
+        versionName '0.8.20220910'
         multiDexEnabled true
         //设置room的Schema的位置
         javaCompileOptions {
